retry: organize crate
This follows the idioms used by the other crates.
